wizard improve (#19)
We want to eliminate the step that the user has to select the exact match from the drop down menu and then click on the continue button Instead, once the location is scanned and there’s an exact match the location code in the system, take the user automatically to the next screen, to eliminate a step and fasten the process. 
By the way – it is much better if you can also eliminate the drop down menu and also don’t allow user to create new location from here.