improved SO mode (#21)
The sale order mode is dictated by the customer sale order mode we preset on the customer card.
But Sometimes there is a special need for the user to change the order mode from whatever it is to something else.
We need a way to add some “override sale order mode” field that user can select from dropdown menu a different order mode that will not effect the customer card and or the rest of the orders, just the single order they have changed it on. 
In that case, if the user selected to override the order mode this sale only will change its order mode and will not depend on the customer order mode as we set on the customer card.
You can create a checkbox field that read: “Manual Order Mode”, if this is checked there is going to be another filed that is showing that will say “Override SOM” and the user can select from a drop down the options (DTC, MP, WS,SC).
If the check box is checked, then the field OVERRIDE SOM will be required and then the order sale mode will be determine by this instead of the current field you have.