Import so (#23)
Some of our customers want us to ship many small orders to their members. The way they do that is by sending to us a CSV file that contains all of the orders information.

Each customer have a little different CSV file and even from time to time they are changing the format a little bit. 

In this Tab I have created all the necessary fields that we will need to include in the import. And the idea is that in each one of these CSV IMPORT tab’s fields I will write the name of the header of the Colum in their file.