Zip code and shipping rules (#25)
1) In America, some states’s zipcodes are
Starting with a “0” and excel 
Automatically removes the “0”.
so some zipcodes on the CSV file will be
With only 4 digits.
For example:
Zipcode: 10020 will be 10020 (no change)
Zipcode: 7820 will be 07820

2) Based on all order lines we need to have a formula that calculates here the total weight in KG.

3) On the customer record I have set a related field that can be set for the customer.
That way the employees never have to remember the shipping rule. So all orders that will be made for this customer will follow the shipping rules (depending on the Sale order mode) 
Based on the customer and The sale order mode.
We need to show the Order shipping rule that we set in the customer card.
And the shipping Rule Code is the field “ShipStation Code” from the shipping rules module.

4) There is only one small complication that we need to add to slide #5.
Some Rules will have a checked box of conditional weight.
That means that if the total weight of the order is above the weight, I wrote on the shipping rule card then the rule that the order will use will be the Rule of the field “use this rule”. And will reflect on the screen I showed on slide #5.
If items are added to the sale or removed and the total weight of the sale order is going to change, then the rule will get adjusted automatically.
This is the way I have done this method here. However, you can either do this or suggest a different way.